Refactor table components
Refactor table header and table body into their own components. Make the
column headers stateful by building out an array with the elements
contained being the column header labels received from the CalTech API.

This is necessary as it will allow for drag & drop operations
re-ordering the column header array which then in turn will impact the
building of the table body.
